refactor: decouple setup/ runtime classes from production artifact dependency
Move runtime-critical classes out of setup/ into Framework so that
setup/ can be excluded from production deployments without breaking
bin/magento.

Changes:
- Move SchemaSetupInterface impl to Framework\Setup\SchemaSetup
- Move ModuleDataSetupInterface impl to Framework\Setup\ModuleDataSetup
- Move SetupCache to Framework\Setup\SetupCache
- Move AbstractSetupCommand to Framework\Console\Command
- Update app/etc/di.xml preferences to point to Framework classes
- Update Backend maintenance commands to use Framework namespace
- Make Cli.php tolerant of missing setup/ directory (file_exists guard)
- Old setup/ classes now extend Framework ones (backwards compatible)

Resolves: https://github.com/magento/magento2/issues/40695
